requirements.txt
torch==2.2.2+cpu -f https://download.pytorch.org/whl/cpu/torch_stable.html
torchvision==0.17.2+cpu -f https://download.pytorch.org/whl/cpu/torch_stable.html
Pillow==10.2.0
requests==2.28.2
Flask==3.0.2
boto3==1.34.75
